Bug Fix:
Previously, when resize_image was called, it would re-read the image
from disk (adding the amount of shrink, and setting :sequential to true).
This caused previous manipulations to be erased; this was especially
problematic when trying to crop a photo using extract area.

This may not be the best fix, but it works. I was unable to find a method
to set :sequential to true in ruby-vips; perhaps I just didnt look hard
enough.