Delegate libvips processing to ImageProcessing gem
ImageProcessing gem is a wrapper gem around MiniMagick and libvips,
which implements common resizing macros. Instead of the carrierwave-vips
project having to maintain those macros, it can use the ImageProcessing
to do the heavy lifting. The ImageProcessing gem will constantly get
updated with best practices for image processing, so any bug fixes and
performance improvements there will automatically benefit
carrierwave-vips.

The ImageProcessing gem automatically pulls in ruby-vips ">= 2.0.10" as
a dependency, so we remove it from the gemspec. ImageProcessing uses the
improved vips_thumbnail(), so it depends on libvips 8.6+.

This PR additionally adds the `#resize_to_pad` method.